% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.cache_up_to_date}
\alias{.cache_up_to_date}
\title{Check if the Cache is Up to Date}
\usage{
.cache_up_to_date(cache_filename, dump_date)
}
\arguments{
\item{cache_filename}{Character scalar; the full path to the (possible)
cache.}

\item{dump_date}{POSIXct scalar; the date against which the cache should be
compared.}
}
\value{
Logical scalar indicating whether the cache exists and is at least as
new as the latest dump.
}
\description{
Check if the Cache is Up to Date
}
\keyword{internal}
