% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache_wikitext}
\alias{.cache_wikitext}
\title{Return a Cached wikitext_en}
\usage{
.cache_wikitext(cache_dir = wikimorphemes_cache_dir())
}
\arguments{
\item{cache_dir}{Character scalar; a path to a cache directory.}
}
\value{
A wikitext_en tibble.
}
\description{
Return a Cached wikitext_en
}
\keyword{internal}
