% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.check_alt_spelling}
\alias{.check_alt_spelling}
\title{Check for Alternative Spelling}
\usage{
.check_alt_spelling(
  wt,
  word,
  sight_words,
  use_lookup,
  current_depth,
  max_depth,
  stop_at
)
}
\arguments{
\item{wt}{Character; wikitext of a word}

\item{word}{Character; the word, to make sure this isn't a loop.}

\item{sight_words}{Character vector; words to \emph{not} break down further.
Defaults to the included \code{sight_words} list; to include no sight words,
pass in an empty character vector to this parameter.}

\item{use_lookup}{Logical; whether to use a cached lookup table (if
available) or always process the word from scratch. If the word is not
available in the lookup, processing (and likely a call to the Wiktionary
API) will still occur. You might want to set this value to FALSE if you've
made recent edits to Wiktionary or otherwise want to see if something has
changed recently.}

\item{current_depth}{Integer; current recursion depth.}

\item{max_depth}{Integer; maximum recursion depth.}

\item{stop_at}{Character; a word to stop processing at. Used to help prevent
loops.}
}
\value{
Character; the first-level of the alternate word or words.
}
\description{
Checks the wikitext for an alternative spelling.
}
\keyword{internal}
