% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.check_alt_spelling_wt}
\alias{.check_alt_spelling_wt}
\title{Check for Simpler Alternative Spelling}
\usage{
.check_alt_spelling_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word split up into component words.
}
\description{
Checks the wikitext for an alternative spelling that breaks the word up into
smaller pieces (e.g. "passerby" -> "passer-by").
}
\keyword{internal}
