% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.check_inflection_ending}
\alias{.check_inflection_ending}
\title{Check that Inflection is Where it is Expected}
\usage{
.check_inflection_ending(original_word, base_word, inflection)
}
\arguments{
\item{original_word}{Character; the word before inflectional breakdown.}

\item{base_word}{Character; uninflected word.}

\item{inflection}{Character; the (putative) inflectional ending.}
}
\value{
TRUE if the word contains the candidate inflection as expected.
}
\description{
In most regular English cases, the word should end with the inflection. But
there are a few exceptions of the "passersby" class, which we want to allow.
}
\keyword{internal}
