% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.check_misspelling_wt}
\alias{.check_misspelling_wt}
\title{Check for Misspelling}
\usage{
.check_misspelling_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the better spelling of the word.
}
\description{
Checks the wikitext to see if this is a misspelling of another word.
}
\keyword{internal}
