% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_nonexplosive_word}
\alias{.check_nonexplosive_word}
\title{Check that Word Breakdown doesn't Contain Original Word}
\usage{
.check_nonexplosive_word(original_word, ...)
}
\arguments{
\item{original_word}{Character; the word before breakdown.}

\item{...}{Character pieces of the word after breakdown.}
}
\value{
TRUE if the word breakdown does not contain the original word as a
piece.
}
\description{
Check that Word Breakdown doesn't Contain Original Word
}
\keyword{internal}
