% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_reconstructed_word}
\alias{.check_reconstructed_word}
\title{Check that Word Breakdown is Close to Original Word}
\usage{
.check_reconstructed_word(original_word, ...)
}
\arguments{
\item{original_word}{Character; the word before breakdown.}

\item{...}{Character arguments that will be pasted together to reconstruct
word.}
}
\value{
TRUE if the reconstructed word is "close enough" to the original.
}
\description{
Check that Word Breakdown is Close to Original Word
}
\keyword{internal}
