% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clean_wiktionary_category_list}
\alias{.clean_wiktionary_category_list}
\title{Get the Members of a Category Return}
\usage{
.clean_wiktionary_category_list(query_return)
}
\arguments{
\item{query_return}{List; the return from
\code{\link[WikipediR]{pages_in_category}}.}
}
\value{
A tibble of results.
}
\description{
Get the Members of a Category Return
}
\keyword{internal}
