% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.clean_word_reference}
\alias{.clean_word_reference}
\title{Strip off Deep Links}
\usage{
.clean_word_reference(word)
}
\arguments{
\item{word}{The word (extracted from a template, probably) to clean.}
}
\value{
Just the word.
}
\description{
Words can link to specific sections. Any time we return a word, we really
just care about the word, not the deeper link.
}
\keyword{internal}
