% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.correct_names}
\alias{.correct_names}
\title{Deal with Piece Names}
\usage{
.correct_names(all_pieces)
}
\arguments{
\item{all_pieces}{A morpheme breakdown, which will likely be a 2-level named
list.}
}
\value{
The morpheme breakdown as a character vector with the desired names.
}
\description{
Deal with Piece Names
}
\keyword{internal}
