% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.create_wikitext_en}
\alias{.create_wikitext_en}
\title{Process the Latest Wiktionary Dump File}
\usage{
.create_wikitext_en()
}
\value{
A logical scalar (invisibly) indicating whether the data was updated.
}
\description{
Using this package can result in a lot of hits to the wiktionary API. To
reduce that burden, this function downloads and processes the entire
wiktionary dump, to create a local version of the English words from the
latest dump.
}
\details{
This function requires a lot of RAM and is very slow. It took about 4 hrs to
complete on a machine with 64GB of RAM.

Note that the data generated in this process is copyright the Wikimedia
Foundation, Inc, with a creative commons BY-SA license. See
\url{https://foundation.wikimedia.org/wiki/Terms_of_Use/en} for more details.
}
\keyword{internal}
