% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.create_word_lists}
\alias{.create_word_lists}
\title{Generate the List of Wiktionary Words}
\usage{
.create_word_lists()
}
\value{
TRUE (invisibly) on success.
}
\description{
This is simply the unique words from \code{\link{.cache_wikitext}}, sorted.
}
\keyword{internal}
