% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.detect_irregular_wt}
\alias{.detect_irregular_wt}
\title{Determine Membership in Irregular Category from Wikitext}
\usage{
.detect_irregular_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext resulting from a call to .fetch_english_word.}
}
\value{
TRUE if the word in question is in an irregular word category.
}
\description{
Determine Membership in Irregular Category from Wikitext
}
\keyword{internal}
