% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.download_latest_dump}
\alias{.download_latest_dump}
\title{Download the Latest Wiktionary Dump}
\usage{
.download_latest_dump()
}
\value{
The path to the dump tempfile.
}
\description{
Download the Latest Wiktionary Dump
}
\keyword{internal}
