% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_english_wt}
\alias{.extract_english_wt}
\title{Extract the English Section of a Wikitext Entry}
\usage{
.extract_english_wt(wt)
}
\arguments{
\item{wt}{The word's page, in wikitext format (returned by
\code{\link{.fetch_word}}).}
}
\value{
The English portion of the wikitext.
}
\description{
Pull just the English section out of a wikitext entry.
}
\keyword{internal}
