% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.extract_relevant_english_wt}
\alias{.extract_relevant_english_wt}
\title{Extract the Relevant English Sections of a Wikitext Entry}
\usage{
.extract_relevant_english_wt(wt)
}
\arguments{
\item{wt}{The word's page, in wikitext format (returned by
\code{\link{.fetch_word}}).}
}
\value{
The relevant sections from the English portion of the wikitext.
}
\description{
Pull just the English section out of a wikitext entry and drop things like
extra etymologies or anagrams.
}
\keyword{internal}
