% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.fetch_english_word}
\alias{.fetch_english_word}
\title{Fetch the English Section of a Wiktionary Word Page}
\usage{
.fetch_english_word(word)
}
\arguments{
\item{word}{The word to fetch.}
}
\value{
Character; the "English" section of the word's page, in wikitext
format.
}
\description{
Fetch the English Section of a Wiktionary Word Page
}
\keyword{internal}
