% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.fetch_redirect}
\alias{.fetch_redirect}
\title{Fetch the Redirect Content for a Word}
\usage{
.fetch_redirect(wt)
}
\arguments{
\item{wt}{Wikitext that begins with "#REDIRECT".}
}
\value{
The wt of the target word.
}
\description{
Some words don't technically have their own entry, but instead are a redirect
to another word.
}
\keyword{internal}
