% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_word}
\alias{.fetch_word}
\title{Fetch the Content of a Wiktionary Word Page}
\usage{
.fetch_word(word)
}
\arguments{
\item{word}{The word to fetch.}
}
\value{
Character; the word's page, in wikitext format.
}
\description{
Fetch the Content of a Wiktionary Word Page
}
\keyword{internal}
