% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.find_page_info}
\alias{.find_page_info}
\title{Create a Tibble of Page Info}
\usage{
.find_page_info(dump_filename)
}
\arguments{
\item{dump_filename}{Character scalar; the full path to a dump tempfile.}
}
\value{
A tibble of rows with \code{<page>} tags, and information about the
spacing between them.
}
\description{
Create a Tibble of Page Info
}
\keyword{internal}
