% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphens.R
\name{.fix_hyphens}
\alias{.fix_hyphens}
\title{Correct Hyphens in Processed Words}
\usage{
.fix_hyphens(processed_word)
}
\arguments{
\item{processed_word}{Character vector; the breakdown of the word.}
}
\value{
The processed words with proper hyphens.
}
\description{
Sometimes we gain or lose hyphens in various processes. Standardize them.
}
\keyword{internal}
