% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.get_latest_dump_date}
\alias{.get_latest_dump_date}
\title{Check wiktionary for the Date of the Latest Dump}
\usage{
.get_latest_dump_date()
}
\value{
The dump date as a POSIXct object, in GMT.
}
\description{
Check wiktionary for the Date of the Latest Dump
}
\keyword{internal}
