% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.list_pages_in_category}
\alias{.list_pages_in_category}
\title{List All Pages in a Category}
\usage{
.list_pages_in_category(categories, type = c("page", "subcat", "file"))
}
\arguments{
\item{categories}{The names of the categories you want to gather information for.}

\item{type}{The type of member you're interested in returning;
options are any permutation of "page" (pages), "subcat" (subcategories) and "file" (files).}
}
\value{
A tibble of results.
}
\description{
Wiktionary only returns 500 results at a time. Many categories have far more
than 500 results. Use this function to get them all.
}
\keyword{internal}
