% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_template_pattern}
\alias{.make_template_pattern}
\title{Construct Regex for Wiktionary Templates}
\usage{
.make_template_pattern(type)
}
\arguments{
\item{type}{Character; specifies which template to construct.}
}
\value{
regular expression for detecting the given type.
}
\description{
Currently works for templates of the "affix" type.
}
\keyword{internal}
