% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.populate_env_lookup}
\alias{.populate_env_lookup}
\title{Populate the Environment Lookup with Saved Results}
\usage{
.populate_env_lookup()
}
\value{
Logical, invisibly; whether a cached lookup exists in that location.
}
\description{
Populate the Environment Lookup with Saved Results
}
\keyword{internal}
