% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.process_wiktionary_dump}
\alias{.process_wiktionary_dump}
\title{Process the Latest Wiktionary Dump File}
\usage{
.process_wiktionary_dump(cache_dir = wikimorphemes_cache_dir())
}
\arguments{
\item{cache_dir}{Character scalar; a path to a cache directory.}
}
\value{
The latest dump date, invisibly.
}
\description{
Using this package can result in a lot of hits to the wiktionary API. To
reduce that burden, this function downloads and processes the entire
wiktionary dump, to create a local version of the English words from the
latest dump.
}
\details{
This function requires a lot of RAM and is very slow.
}
\keyword{internal}
