% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.process_word_recursive}
\alias{.process_word_recursive}
\title{Split a Word into Pieces}
\usage{
.process_word_recursive(
  word,
  use_lookup = TRUE,
  cache_dir = wikimorphemes_cache_dir(),
  current_depth = 1,
  max_depth = 30
)
}
\arguments{
\item{word}{Character; a word to process.}

\item{use_lookup}{Logical; should we use a cached lookup table if available,
or go straight to the Wiktionary API. You might want to set this value to
FALSE if you've made recent edits to Wiktionary or otherwise want to see if
something has changed recently.}

\item{cache_dir}{Character scalar; a path to a cache directory.}

\item{current_depth}{Integer; current recursion depth.}

\item{max_depth}{Integer; maximum recursion depth.}
}
\value{
Character; the word split into pieces.
}
\description{
Recursively splits a word into constituent pieces, based on Wiktionary
annotations. There are two main categories of word pieces used: inflections
(standard verb/noun/comparative adjective forms, defined in practice here as
endings identified in Wiktionary by inflectional function without reference
to the actual form of the ending) and morphemes (typically denoted in
Wiktionary by etymology templates).
}
\keyword{internal}
