% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.recombine_sections}
\alias{.recombine_sections}
\title{Recombine Sections into Wikitext}
\usage{
.recombine_sections(sections, depth)
}
\arguments{
\item{sections}{A named character vector of sections returned by
\code{\link{.split_sections}}.}

\item{depth}{The depth at which these sections were extracted.}
}
\value{
A character scalar of the recombined sections.
}
\description{
Recombine Sections into Wikitext
}
\keyword{internal}
