% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_any_character_except}
\alias{.regex_any_character_except}
\title{Helper to Generate Regex For Character Exceptions}
\usage{
.regex_any_character_except(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to not allow in the regex.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex For Character Exceptions
}
\keyword{internal}
