% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_negative_lookahead}
\alias{.regex_negative_lookahead}
\title{Helper to Generate Regex Negative Look Ahead}
\usage{
.regex_negative_lookahead(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to look for (and make sure it doesn't occur) after the "main" regex.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex Negative Look Ahead
}
\keyword{internal}
