% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_one_or_more}
\alias{.regex_one_or_more}
\title{Helper to Generate Regex One or More}
\usage{
.regex_one_or_more(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to check for at least one of.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex One or More
}
\keyword{internal}
