% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_or}
\alias{.regex_or}
\title{Helper to Generate Regex Or}
\usage{
.regex_or(strings)
}
\arguments{
\item{string}{Character; strings to concatenate.}
}
\value{
A length-1 string.
}
\description{
Helper to Generate Regex Or
}
\keyword{internal}
