% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_positive_lookahead}
\alias{.regex_positive_lookahead}
\title{Helper to Generate Regex Positive Look Ahead}
\usage{
.regex_positive_lookahead(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to look for after the "main" regex.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex Positive Look Ahead
}
\keyword{internal}
