% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_positive_lookbehind}
\alias{.regex_positive_lookbehind}
\title{Helper to Generate Regex Positive Look Behind}
\usage{
.regex_positive_lookbehind(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to look for behind the "main" regex.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex Positive Look Behind
}
\keyword{internal}
