% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{.regex_zero_or_more}
\alias{.regex_zero_or_more}
\title{Helper to Generate Regex Zero or More}
\usage{
.regex_zero_or_more(string)
}
\arguments{
\item{string}{Character; a string or strings to modify with regex. The thing
to allow to exist.}
}
\value{
A modified string.
}
\description{
Helper to Generate Regex Zero or More
}
\keyword{internal}
