% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_affixes_wt}
\alias{.split_affixes_wt}
\title{Split Affixes from a Word}
\usage{
.split_affixes_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word with affixes split off.
}
\description{
Split Affixes from a Word
}
\keyword{internal}
