% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_compounds_wt}
\alias{.split_compounds_wt}
\title{Split Compound Word}
\usage{
.split_compounds_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word split up into component words.
}
\description{
Split Compound Word
}
\keyword{internal}
