% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_confixes_wt}
\alias{.split_confixes_wt}
\title{Split Confixes from a Word}
\usage{
.split_confixes_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word with confixes split off.
}
\description{
In Wiktionary, words with both a suffix and prefix are tagged as "confix".
Also used when there is no base word (only prefix + suffix)
}
\keyword{internal}
