% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_contractions}
\alias{.split_contractions}
\title{Split Contractions}
\usage{
.split_contractions(wt, word)
}
\arguments{
\item{wt}{Character; wikitext of a word}

\item{word}{Character; the word.}
}
\value{
Character; the word split up into component words.
}
\description{
Split Contractions
}
\keyword{internal}
