% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_contractions_wt}
\alias{.split_contractions_wt}
\title{Split Contractions}
\usage{
.split_contractions_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word split up into component words.
}
\description{
Split Contractions
}
\keyword{internal}
