% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_inflections}
\alias{.split_inflections}
\title{Split Standard Verb, Noun, and Adjective Endings of a Word}
\usage{
.split_inflections(english_content, word)
}
\arguments{
\item{english_content}{The return from \code{\link{.fetch_english_word}}.}

\item{word}{Character; a word to process.}
}
\value{
Character; the word with standard endings split off.
}
\description{
Split Standard Verb, Noun, and Adjective Endings of a Word
}
\keyword{internal}
