% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_morphemes}
\alias{.split_morphemes}
\title{Split Word into Morphemes}
\usage{
.split_morphemes(word)
}
\arguments{
\item{word}{Character; a word to process.}
}
\value{
Character; the word split into morphemes.
}
\description{
Split Word into Morphemes
}
\keyword{internal}
