% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_on_breaks}
\alias{.split_on_breaks}
\title{Split on Break Characters}
\usage{
.split_on_breaks(word)
}
\arguments{
\item{word}{The word to potentially break down.}
}
\value{
The breakdown, if there are break characters.
}
\description{
This is a fall-through if nothing else breaks the word down.
}
\keyword{internal}
