% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_possessive}
\alias{.split_possessive}
\title{Split the Possessive s}
\usage{
.split_possessive(word)
}
\arguments{
\item{word}{The word to split.}
}
\value{
The word, split into word + -'s if it is possessive.
}
\description{
This is probably not exhaustive, and could potentially have some false
positives, but it should only happen when templates fail us so I think it
will usually be correct.
}
\keyword{internal}
