% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_prefixes_wt}
\alias{.split_prefixes_wt}
\title{Split Prefixes from a Word}
\usage{
.split_prefixes_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word with prefixes split off.
}
\description{
Split Prefixes from a Word
}
\keyword{internal}
