% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{.split_suffixes_wt}
\alias{.split_suffixes_wt}
\title{Split Suffixes from a Word}
\usage{
.split_suffixes_wt(wt)
}
\arguments{
\item{wt}{Character; wikitext of a word}
}
\value{
Character; the word with suffixes split off.
}
\description{
Split Suffixes from a Word
}
\keyword{internal}
