% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{.update_env_lookup}
\alias{.update_env_lookup}
\title{Update the Environment Lookup}
\usage{
.update_env_lookup(word, morphemes, use_lookup)
}
\arguments{
\item{word}{The word that has been processed.}

\item{morphemes}{The processed morphemes to save in the cache.}

\item{use_lookup}{Logical; whether to use a cached lookup table (if
available) or always process the word from scratch. If the word is not
available in the lookup, processing (and likely a call to the Wiktionary
API) will still occur. You might want to set this value to FALSE if you've
made recent edits to Wiktionary or otherwise want to see if something has
changed recently.}
}
\value{
The morphemes, with any post-processing that's needed on all
morphemes (right now, removing duplicate hyphens).
}
\description{
When we look up a word and that word isn't in the existing lookup, we add it
to the lookup.
}
\keyword{internal}
