% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{process_word_recursive}
\alias{process_word_recursive}
\title{Split a Word into Pieces}
\usage{
process_word_recursive(word)
}
\arguments{
\item{word}{Character; a word to process.}
}
\value{
Character; the word split into pieces.
}
\description{
Recursively splits a word into constituent pieces, based on Wiktionary
annotations.There are two main categories of word pieces used: inflections
(standard verb/noun/comparative adjective forms, defined in practice here as
endings identified in Wiktionary by inflectional function without reference
to the actual form of the ending) and morphemes (typically denoted in
Wiktionary by etymology templates).
}
