% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.data}
\alias{.env}
\alias{\%>\%}
\alias{memoise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{memoise}{\code{\link[memoise]{memoise}}}

  \item{rlang}{\code{\link[rlang:tidyeval-data]{.data}}, \code{\link[rlang:tidyeval-data]{.env}}}
}}

