% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{reset_lookup}
\alias{reset_lookup}
\title{Reset the Lookup Cache}
\usage{
reset_lookup()
}
\value{
TRUE (invisibly) on success.
}
\description{
Sometimes it is useful to reset the lookup cache, for example if you have
made a change on wiktionary and wish to use the newest result. This function
resets the in-memory lookup cache. This is likely a very rare situation.
}
\examples{
reset_lookup()
}
