% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{save_wikimorphemes_lookup}
\alias{save_wikimorphemes_lookup}
\title{Save the Lookup Cache to Disk}
\usage{
save_wikimorphemes_lookup(cache_dir = wikimorphemes_cache_dir())
}
\arguments{
\item{cache_dir}{Character scalar; a path to a cache directory.}
}
\value{
TRUE, invisibly.
}
\description{
Unless you specify otherwise, wikimorphemes produces a cache in RAM of the
words you have parsed. Use this function to update (or create) an on-disk
version of this lookup table.
}
