% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wikimorphemes_cache_dir}
\alias{wikimorphemes_cache_dir}
\title{Wikimorphemes Cache Directory}
\usage{
wikimorphemes_cache_dir()
}
\value{
A character vector with the normalized path to the cache.
}
\description{
The wikimorphemes cache directory is a platform- and user-specific path where
wikimorpheme saves caches (such as a downloaded lookup). You can override the
default location in a few ways:
\itemize{
\item{Option: \code{wikimorphemes.dir}}{Use
\code{\link{set_wikimorphemes_cache_dir}} to set a specific cache directory
for this session}
\item{Environment: \code{WIKIMORPHEMES_CACHE_DIR}}{Set this environment
variable to specify a wikimorphemes cache directory for all sessions.}
\item{Environment: \code{R_USER_CACHE_DIR}}{Set this environment variable
to specify a cache directory root for all packages that use the caching
system.}
}
}
\examples{
wikimorphemes_cache_dir()
}
