% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wikimorphemes_lookup}
\alias{wikimorphemes_lookup}
\title{Load the Current Lookup}
\usage{
wikimorphemes_lookup()
}
\value{
A tibble with columns word, morphemes (a list column with a character
vector of morphemes per word), and n_morphemes (an integer count of
morphemes in the morphemes column). Note that words less than 4 characters
are never broken into morphemes by this package so they do not appear in
the lookup. If you do not have a lookup, this function will return NULL.
}
\description{
During use, Wikimorphemes builds an in-memory lookup table of words and
morphemes. This table is used by \code{\link{process_word}}, but it might
also be useful for other uses.
}
\examples{
nrow(wikimorphemes_lookup())
head(wikimorphemes_lookup())
}
