% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wiktionary_word_list}
\alias{wiktionary_word_list}
\title{Load the List of Words in Wiktionary}
\usage{
wiktionary_word_list()
}
\value{
A character vector with the unique words available in the English
Wiktionary.
}
\description{
Each time we process a dump from Wiktionary, we also save a separate list of
Wiktionary words. If you have already downloaded the word list, it will not
download again. Right now you need to manually delete the word list file to
remove it from your cache. Note: this will download a 4.12 MB file if you do
not already have it in your cache.
}
