% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache_wikitext}
\alias{.cache_wikitext}
\title{Return a Cached wikitext_en}
\usage{
.cache_wikitext()
}
\value{
A wikitext_en tibble. Note that this tibble is copyright the
Wikimedia Foundation, Inc, with a creative commons BY-SA license. See
\url{https://foundation.wikimedia.org/wiki/Terms_of_Use/en} for more
details.
}
\description{
Return a Cached wikitext_en
}
\keyword{internal}
