% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{.validate_cache_dir_write}
\alias{.validate_cache_dir_write}
\title{Make Sure the Cache Dir is Writable}
\usage{
.validate_cache_dir_write()
}
\value{
The cache_dir, if it is writable.
}
\description{
Make Sure the Cache Dir is Writable
}
\keyword{internal}
