% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_splitters.R
\name{process_word}
\alias{process_word}
\title{Split a Word into Pieces}
\usage{
process_word(word)
}
\arguments{
\item{word}{Character; a word to process.}
}
\value{
Character; the word split into pieces.
}
\description{
Splits a word into constituent pieces. Wrapper around
\code{\link{process_word_recursive}}, with some postprocessing.
}
