% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_wikimorphemes_cache_dir}
\alias{set_wikimorphemes_cache_dir}
\title{Set a Cache Directory for Wikimorphemes}
\usage{
set_wikimorphemes_cache_dir(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character scalar; a path to a cache directory.}
}
\value{
A normalized path to a cache directory. The directory is created if
the user has write access and the directory does not exist.
}
\description{
By default, this package uses a cache directory chosen using
/code{/link[dlr]{app_cache_dir}}. If you wish to set a different directory,
call this function. This function is also called when the package loads. If
you wish to always set a different cache directory, you can set the
environment variable WIKIMORPHEMES_CACHE_DIR or the option wikimorphemes.dir.
Note that this function sets an option when called.
}
