% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wikimorphemes_cache_dir}
\alias{wikimorphemes_cache_dir}
\title{Choose a Cache Directory for Wikimorphemes}
\usage{
wikimorphemes_cache_dir(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character scalar; a path to a cache directory.}
}
\value{
A normalized path to a cache directory. The directory is created if
the user has write access and the directory does not exist. Note that this
function also sets an option, "wikimorphemes.dir", to simplify subsequent
checks.
}
\description{
It is often useful to work with a local cache for this package. This function
helps to locate the cache directory. In most cases, you can just use the
default directory.
}
